♻️  refactor: external 모듈 분리에 따른 변경사항 반영 #124
